restructured the package to have barrel files
