+++ src/core/index.ts
@@ -1,0 +1,13 @@
+export { BlockFetcher } from './block-fetcher';
+export type { BlockFetcherConfig } from './block-fetcher';
+
+export { Exporter } from './exporter';
+
+export { MediaHandler } from './media-handler';
+export { DownloadStrategy } from './media-handler/strategies/download';
+export { UploadStrategy } from './media-handler/strategies/upload';
+
+export { PageReferenceHandler } from './page-ref-handler';
+export type { PageRefConfig } from './page-ref-handler';
+
+export { BaseRendererPlugin } from './renderer';
