+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+export {
+  MediaManifestManager,
+  PageReferenceManifestManager,
+} from './manifest-manager';
+
+export { PageReferenceManifestBuilder } from './page-ref-builder';
